refactor(agents): isolate harness-specific subagent code behind one registry
Each harness now lives in its own file under terminal-agents/subagent-
harnesses (claude, codex, generic) built with defineSubagentHarness, which
fills every member with the shared default so an entry only has to bring a
parser. The registry owns the stop-event rule, the parent-session check,
path resolution, and description lookup; the transcript module keeps only
the generic entry type, tail reader, and input summarizer.

The store owns the subagent pipeline (recordSubagentHook,
getSubagentTranscript) so both routers shrink to validation and broadcast,
and the trust check for hook-supplied paths runs where the data is owned.

Renderer: one search-param codec (buildSubagentSearch/parseSubagentSearch)
replaces five spellings of the deep-link fields and validates the agent id;
focusOrOpenPane is shared by the page and subagent openers; the pane kind
is a constant; the transcript row uses the host's exported entry type.

Shell: one dispatch_to_host function serves both the agent and subagent
payloads (marker v13).
