+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/utils/workspace-navigation.ts
@@ -1,3 +1,7 @@
+import {
+	AGENT_IDENTITY_IDS,
+	type AgentIdentityId,
+} from "@superset/shared/agent-catalog";
 import type {
 	NavigateOptions,
 	UseNavigateResult,
@@ -8,13 +12,63 @@ export interface WorkspaceSearchParams {
 	paneId?: string;
 }
 
-export interface V2WorkspaceSearchParams {
-	terminalId?: string;
-	focusRequestId?: string;
+/**
+ * Deep link into a subagent's transcript pane. The sidebar cannot reach a
+ * workspace's pane store, so it navigates with these and the workspace
+ * page opens the pane (see useConsumeSubagentLink).
+ */
+export interface SubagentLinkParams {
+	terminalId: string;
+	subagentId: string;
+	agentId: AgentIdentityId;
+	agentType?: string;
+}
+
+export interface SubagentLinkSearchParams {
 	subagentTerminalId?: string;
 	subagentId?: string;
 	subagentAgentId?: string;
 	subagentType?: string;
+}
+
+export function buildSubagentSearch(
+	link: SubagentLinkParams,
+): SubagentLinkSearchParams {
+	return {
+		subagentTerminalId: link.terminalId,
+		subagentId: link.subagentId,
+		subagentAgentId: link.agentId,
+		...(link.agentType ? { subagentType: link.agentType } : {}),
+	};
+}
+
+function nonEmptyString(value: unknown): string | undefined {
+	return typeof value === "string" && value.length > 0 ? value : undefined;
+}
+
+/** The subagent link in a raw search object, or undefined when absent or malformed. */
+export function parseSubagentSearch(
+	raw: Record<string, unknown>,
+): SubagentLinkParams | undefined {
+	const terminalId = nonEmptyString(raw.subagentTerminalId);
+	const subagentId = nonEmptyString(raw.subagentId);
+	const agentId = nonEmptyString(raw.subagentAgentId);
+	if (!terminalId || !subagentId || !agentId) return undefined;
+	if (!(AGENT_IDENTITY_IDS as readonly string[]).includes(agentId)) {
+		return undefined;
+	}
+	const agentType = nonEmptyString(raw.subagentType);
+	return {
+		terminalId,
+		subagentId,
+		agentId: agentId as AgentIdentityId,
+		...(agentType ? { agentType } : {}),
+	};
+}
+
+export interface V2WorkspaceSearchParams extends SubagentLinkSearchParams {
+	terminalId?: string;
+	focusRequestId?: string;
 	openUrl?: string;
 	openUrlTarget?: "current-tab" | "new-tab";
 	openUrlRequestId?: string;
