+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/useConsumeSubagentLink/useConsumeSubagentLink.ts
@@ -1,6 +1,6 @@
 import type { WorkspaceStore } from "@superset/panes";
-import type { AgentIdentityId } from "@superset/shared/agent-catalog";
 import { useEffect, useRef } from "react";
+import type { SubagentLinkParams } from "renderer/routes/_authenticated/_dashboard/utils/workspace-navigation";
 import type { StoreApi } from "zustand/vanilla";
 import type { PaneViewerData } from "../../types";
 import { openSubagentPaneInStore } from "../../utils/openSubagentPaneInStore";
@@ -8,47 +8,28 @@ import { openSubagentPaneInStore } from "../../utils/openSubagentPaneInStore";
 interface UseConsumeSubagentLinkArgs {
 	store: StoreApi<WorkspaceStore<PaneViewerData>>;
 	isLayoutReady: boolean;
-	terminalId: string | undefined;
-	subagentId: string | undefined;
-	agentId: string | undefined;
-	agentType: string | undefined;
+	link: SubagentLinkParams | undefined;
 	focusRequestId: string | undefined;
 }
 
 /**
- * Opens the subagent transcript pane named by the workspace's search params
- * (`?subagentTerminalId=…&subagentId=…`), the way the sidebar's agents chip
- * deep-links into a workspace it cannot reach the pane store of. Each
- * request id is consumed once so a re-render does not reopen the pane.
+ * Opens the subagent transcript pane named by the workspace's search params,
+ * the way the sidebar's agents chip deep-links into a workspace it cannot
+ * reach the pane store of. Each request id is consumed once so a re-render
+ * does not reopen the pane.
  */
 export function useConsumeSubagentLink({
 	store,
 	isLayoutReady,
-	terminalId,
-	subagentId,
-	agentId,
-	agentType,
+	link,
 	focusRequestId,
 }: UseConsumeSubagentLinkArgs): void {
 	const consumedRef = useRef<Set<string>>(new Set());
 	useEffect(() => {
-		if (!isLayoutReady || !terminalId || !subagentId || !agentId) return;
-		const key = `${terminalId}:${subagentId}:${focusRequestId ?? ""}`;
+		if (!isLayoutReady || !link) return;
+		const key = `${link.terminalId}:${link.subagentId}:${focusRequestId ?? ""}`;
 		if (consumedRef.current.has(key)) return;
 		consumedRef.current.add(key);
-		openSubagentPaneInStore(store, {
-			terminalId,
-			subagentId,
-			agentId: agentId as AgentIdentityId,
-			...(agentType ? { agentType } : {}),
-		});
-	}, [
-		store,
-		isLayoutReady,
-		terminalId,
-		subagentId,
-		agentId,
-		agentType,
-		focusRequestId,
-	]);
+		openSubagentPaneInStore(store, link);
+	}, [store, isLayoutReady, link, focusRequestId]);
 }
