+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/utils/focusOrOpenPane/index.ts
@@ -1,0 +1,1 @@
+export { focusOrOpenPane } from "./focusOrOpenPane";
