+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/utils/openSubagentPaneInStore/openSubagentPaneInStore.ts
@@ -1,6 +1,11 @@
 import type { WorkspaceStore } from "@superset/panes";
 import type { StoreApi } from "zustand/vanilla";
-import type { PaneViewerData, SubagentPaneData } from "../../types";
+import {
+	type PaneViewerData,
+	SUBAGENT_PANE_KIND,
+	type SubagentPaneData,
+} from "../../types";
+import { focusOrOpenPane } from "../focusOrOpenPane";
 
 /**
  * Focus the pane showing this subagent's transcript, creating one when none
@@ -10,21 +15,12 @@ export function openSubagentPaneInStore(
 	store: StoreApi<WorkspaceStore<PaneViewerData>>,
 	data: SubagentPaneData,
 ): void {
-	const state = store.getState();
-	for (const tab of state.tabs) {
-		for (const pane of Object.values(tab.panes)) {
-			if (pane.kind !== "subagent") continue;
-			const existing = pane.data as SubagentPaneData;
-			if (
-				existing.terminalId !== data.terminalId ||
-				existing.subagentId !== data.subagentId
-			) {
-				continue;
-			}
-			state.setActiveTab(tab.id);
-			state.setActivePane({ tabId: tab.id, paneId: pane.id });
-			return;
-		}
-	}
-	state.openPane({ pane: { kind: "subagent", data: data as PaneViewerData } });
+	focusOrOpenPane<SubagentPaneData>(
+		store,
+		SUBAGENT_PANE_KIND,
+		(pane) =>
+			pane.terminalId === data.terminalId &&
+			pane.subagentId === data.subagentId,
+		data,
+	);
 }
