+++ packages/host-service/src/terminal-agents/subagent-transcript.ts
@@ -26,17 +26,17 @@ export interface SubagentTranscript {
 const MAX_READ_BYTES = 2 * 1024 * 1024;
 const MAX_ENTRY_CHARS = 4000;
 
-function clip(text: string): string {
+export function clip(text: string): string {
 	return text.length > MAX_ENTRY_CHARS
 		? `${text.slice(0, MAX_ENTRY_CHARS)}…`
 		: text;
 }
 
-function isRecord(value: unknown): value is Record<string, unknown> {
+export function isRecord(value: unknown): value is Record<string, unknown> {
 	return typeof value === "object" && value !== null && !Array.isArray(value);
 }
 
-function asString(value: unknown): string {
+export function asString(value: unknown): string {
 	if (typeof value === "string") return value;
 	if (value === undefined || value === null) return "";
 	try {
@@ -46,14 +46,14 @@ function asString(value: unknown): string {
 	}
 }
 
-function parseTimestamp(value: unknown): number | undefined {
+export function parseTimestamp(value: unknown): number | undefined {
 	if (typeof value !== "string") return undefined;
 	const ms = Date.parse(value);
 	return Number.isNaN(ms) ? undefined : ms;
 }
 
-/** Short, human summary of a tool call's input. */
-function summarizeToolInput(input: unknown): string {
+/** Short, human summary of a tool call's input, shared by every harness parser. */
+export function summarizeToolInput(input: unknown): string {
 	if (typeof input === "string") {
 		try {
 			return summarizeToolInput(JSON.parse(input));
@@ -85,224 +85,6 @@ function summarizeToolInput(input: unknown): string {
 		}
 	}
 	return clip(asString(input));
-}
-
-/** Text of a Claude / Anthropic content block list. */
-function blocksText(content: unknown): string {
-	if (typeof content === "string") return content;
-	if (!Array.isArray(content)) return "";
-	return content
-		.map((block) =>
-			isRecord(block) && typeof block.text === "string" ? block.text : "",
-		)
-		.filter(Boolean)
-		.join("\n");
-}
-
-export function parseClaudeSubagentTranscript(
-	text: string,
-): SubagentTranscriptEntry[] {
-	const entries: SubagentTranscriptEntry[] = [];
-	let lineNumber = 0;
-	for (const line of text.split("\n")) {
-		lineNumber += 1;
-		if (!line.trim()) continue;
-		let record: unknown;
-		try {
-			record = JSON.parse(line);
-		} catch {
-			continue;
-		}
-		if (!isRecord(record)) continue;
-		const message = isRecord(record.message) ? record.message : undefined;
-		if (!message) continue;
-		const type = record.type;
-		if (type !== "user" && type !== "assistant") continue;
-		const baseId =
-			typeof record.uuid === "string" ? record.uuid : `line-${lineNumber}`;
-		const timestamp = parseTimestamp(record.timestamp);
-		const content = message.content;
-
-		if (typeof content === "string") {
-			if (content.trim()) {
-				entries.push({
-					id: baseId,
-					kind: type === "assistant" ? "assistant" : "user",
-					text: clip(content),
-					timestamp,
-				});
-			}
-			continue;
-		}
-		if (!Array.isArray(content)) continue;
-		content.forEach((block, index) => {
-			if (!isRecord(block)) return;
-			const id = `${baseId}:${index}`;
-			switch (block.type) {
-				case "text": {
-					const body = typeof block.text === "string" ? block.text.trim() : "";
-					if (!body) return;
-					entries.push({
-						id,
-						kind: type === "assistant" ? "assistant" : "user",
-						text: clip(body),
-						timestamp,
-					});
-					return;
-				}
-				case "thinking": {
-					const body =
-						typeof block.thinking === "string" ? block.thinking.trim() : "";
-					if (!body) return;
-					entries.push({ id, kind: "thinking", text: clip(body), timestamp });
-					return;
-				}
-				case "tool_use":
-					entries.push({
-						id,
-						kind: "tool_call",
-						toolName: typeof block.name === "string" ? block.name : "tool",
-						text: summarizeToolInput(block.input),
-						timestamp,
-					});
-					return;
-				case "tool_result": {
-					const body = blocksText(block.content).trim();
-					entries.push({
-						id,
-						kind: "tool_result",
-						text: clip(body),
-						timestamp,
-					});
-					return;
-				}
-				default:
-					return;
-			}
-		});
-	}
-	return entries;
-}
-
-/** Text of a Codex message content list (`input_text` / `output_text`). */
-function codexContentText(content: unknown): string {
-	if (typeof content === "string") return content;
-	if (!Array.isArray(content)) return "";
-	return content
-		.map((part) =>
-			isRecord(part) && typeof part.text === "string" ? part.text : "",
-		)
-		.filter(Boolean)
-		.join("\n");
-}
-
-export function parseCodexRolloutTranscript(text: string): {
-	entries: SubagentTranscriptEntry[];
-	description?: string;
-} {
-	const entries: SubagentTranscriptEntry[] = [];
-	let description: string | undefined;
-	let line = 0;
-	for (const raw of text.split("\n")) {
-		line += 1;
-		if (!raw.trim()) continue;
-		let record: unknown;
-		try {
-			record = JSON.parse(raw);
-		} catch {
-			continue;
-		}
-		if (!isRecord(record)) continue;
-		const payload = isRecord(record.payload) ? record.payload : undefined;
-		if (!payload) continue;
-		const timestamp = parseTimestamp(record.timestamp);
-		const id = typeof payload.id === "string" ? payload.id : `line-${line}`;
-
-		if (record.type === "session_meta") {
-			const nickname =
-				typeof payload.agent_nickname === "string"
-					? payload.agent_nickname
-					: "";
-			const agentPath =
-				typeof payload.agent_path === "string" ? payload.agent_path : "";
-			description =
-				[nickname, agentPath].filter(Boolean).join(" · ") || undefined;
-			continue;
-		}
-		if (record.type !== "response_item") continue;
-
-		switch (payload.type) {
-			case "message": {
-				const body = codexContentText(payload.content).trim();
-				if (!body) break;
-				if (payload.role === "assistant") {
-					entries.push({ id, kind: "assistant", text: clip(body), timestamp });
-				} else if (payload.role === "user") {
-					entries.push({ id, kind: "user", text: clip(body), timestamp });
-				}
-				break;
-			}
-			case "agent_message": {
-				const body = codexContentText(payload.content).trim();
-				if (!body) break;
-				const author = typeof payload.author === "string" ? payload.author : "";
-				entries.push({
-					id,
-					kind: "user",
-					text: clip(author ? `${author}: ${body}` : body),
-					timestamp,
-				});
-				break;
-			}
-			case "reasoning": {
-				const summary = Array.isArray(payload.summary)
-					? payload.summary
-							.map((part) =>
-								isRecord(part) && typeof part.text === "string"
-									? part.text
-									: "",
-							)
-							.filter(Boolean)
-							.join("\n")
-					: "";
-				if (!summary.trim()) break;
-				entries.push({ id, kind: "thinking", text: clip(summary), timestamp });
-				break;
-			}
-			case "function_call":
-			case "custom_tool_call":
-				entries.push({
-					id,
-					kind: "tool_call",
-					toolName: typeof payload.name === "string" ? payload.name : "tool",
-					text: summarizeToolInput(payload.arguments ?? payload.input),
-					timestamp,
-				});
-				break;
-			case "function_call_output":
-			case "custom_tool_call_output": {
-				let body = asString(payload.output);
-				try {
-					const parsed = JSON.parse(body);
-					if (isRecord(parsed) && typeof parsed.output === "string") {
-						body = parsed.output;
-					}
-				} catch {
-					// plain text output
-				}
-				entries.push({
-					id,
-					kind: "tool_result",
-					text: clip(body.trim()),
-					timestamp,
-				});
-				break;
-			}
-			default:
-				break;
-		}
-	}
-	return { entries, description };
 }
 
 /**
